requirements.txt
dbt-postgres
dbt-snowflake
python-dotenv